refactor Gruntfile
