+++ grunt-helpers/grunt-migrations.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = function (grunt, options) {
+	if (options.enableAngularMigration) {
+		require('../grunt-sections/angular-migration')(grunt, options).addMigration();
+	}
+};
