+++ package.json
@@ -113,7 +113,9 @@
   },
   "scripts": {
     "build": "node_modules/wix-gruntfile/scripts/build.sh",
-    "release": "node_modules/wix-gruntfile/scripts/release.sh"
+    "release": "node_modules/wix-gruntfile/scripts/release.sh",
+    "test": "#tbd",
+    "start": "grunt serve"
   },
   "devDependencies": {}
 }
